fix and refactgor
