+++ zoo/src/test/scala/com/intel/analytics/zoo/pipeline/api/keras/layers/b/TorchMMSpec.scala
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.intel.analytics.zoo.pipeline.api.torch
+package com.intel.analytics.zoo.pipeline.api.keras.layers.b
 
-import com.intel.analytics.bigdl.nn.{Input, Log, MM}
+import com.intel.analytics.bigdl.nn.{Input, Log}
 import com.intel.analytics.bigdl.tensor.Tensor
-import com.intel.analytics.bigdl.utils.{Shape, T, Table}
+import com.intel.analytics.bigdl.utils.{Shape, T}
 import com.intel.analytics.zoo.pipeline.api.autograd.{AutoGrad, TorchMM, Variable}
 import com.intel.analytics.zoo.pipeline.api.keras.models.Model
 import com.intel.analytics.zoo.pipeline.api.keras.serializer.ModuleSerializationTest
@@ -77,7 +77,7 @@ class TorchMMSpec extends FlatSpec with Matchers {
   }
 }
 
-class MMSerialTest extends ModuleSerializationTest {
+class TorchMMSerialTest extends ModuleSerializationTest {
   override def test(): Unit = {
     val mm = TorchMM[Float]().setName("mm_layer")
     val input1 = Tensor[Float](2, 3).apply1(e => Random.nextFloat())
