+++ zoo/src/test/scala/com/intel/analytics/zoo/pipeline/api/keras/serializer/SerializerSpecHelper.scala
@@ -36,7 +36,9 @@ abstract class SerializerSpecHelper extends FlatSpec with Matchers with BeforeAn
 
   private val excluded = Set[String](
     "com.intel.analytics.zoo.pipeline.api.autograd.LambdaTorch",
-    "com.intel.analytics.zoo.pipeline.api.net.TFNet")
+    "com.intel.analytics.zoo.pipeline.api.net.TFNet",
+    "com.intel.analytics.zoo.pipeline.api.keras.layers.b.Recurrent",
+    "com.intel.analytics.zoo.pipeline.api.keras.layers.b.TorchRecurrent")
 
   private val unRegularNameMapping = Map[String, String]()
